Auto-Create Command of sheets

+++ Assets/Capstone/Scripts/Projectile/Projectile.cs
@@ -3,13 +3,13 @@ using UnityEngine;
 public class Projectile : MonoBehaviour
 {
     public GameObject target;
-    public int damage;
+    public float damage;
     public float speed;
     public float destroyTime;
 
     private Rigidbody2D rigid;
 
-    public virtual void Initialize(GameObject target, int damage, float speed, float destroyTime)
+    public virtual void Initialize(GameObject target, float damage, float speed, float destroyTime)
     {
         this.target = target;
         this.damage = damage;
